requirements.txt
datasets==1.15.1
numpy==1.19.2
tqdm==4.62.3
transformers==4.11.3
seqeval==1.2.2
matplotlib
faiss-gpu
